Replace paresPubspec top-level function with Pubspec.parse ctor
